Extracted booking with guests composition into IBookingWithGuestsProvider service
