+++ BookingCalendarApi/Services/IBookingWithGuestsProvider.cs
@@ -1,0 +1,10 @@
+﻿using BookingCalendarApi.Models.Iperbooking.Bookings;
+
+namespace BookingCalendarApi.Services
+{
+    public interface IBookingWithGuestsProvider
+    {
+        public Task FetchAsync(string date);
+        public IEnumerable<AssignedBooking<Models.Iperbooking.Guests.Guest>> Bookings { get; }
+    }
+}
